Remove unrelated Binance utility refactor
